requirements.txt
asgiref==3.8.1
bootstrap4==0.1.0
certifi==2024.8.30
Django==5.1.3
djangorestframework==3.15.2
mysqlclient==2.2.6
pillow==11.0.0
sqlparse==0.5.2
